Extract installation code to the 'installer' module.
Add an `install-ns-webpack` binary.
