+++ bin/install-ns-webpack
@@ -1,0 +1,2 @@
+#!/usr/bin/env node
+var installer = require("../installer");
