+++ bin/install-ns-webpack.cmd
@@ -1,0 +1,1 @@
+@node %~dp0\install-ns-webpack %*
